+++ packages/router/src/bindings/fastify/index.ts
@@ -1,9 +1,8 @@
 import { createRequestContext, jsonifyError } from "@connext/nxtp-utils";
-import { constants, utils } from "ethers";
+import { constants } from "ethers";
 import fastify from "fastify";
 import { register } from "prom-client";
 
-import { signRouterRemoveLiquidityTransactionPayload } from "../../lib/helpers";
 import { getContext } from "../../router";
 import { handleActiveTransactions } from "../contractReader";
 
@@ -26,17 +25,7 @@ import {
 
 export const bindFastify = () =>
   new Promise<void>((res) => {
-    const {
-      wallet,
-      contractWriter,
-      config,
-      logger,
-      contractReader,
-      isRouterContract,
-      txService,
-      chainData,
-      routerAddress,
-    } = getContext();
+    const { wallet, contractWriter, config, logger, contractReader, isRouterContract, routerAddress } = getContext();
 
     const server = fastify();
 
@@ -73,36 +62,16 @@ export const bindFastify = () =>
         try {
           let result;
           if (isRouterContract) {
-            const routerRelayerFeeAsset = utils.getAddress(
-              config.chainConfig[chainId].routerContractRelayerAsset ?? constants.AddressZero,
-            );
-            const relayerFeeAssetDecimal = await txService.getDecimalsForAsset(chainId, routerRelayerFeeAsset);
-            const routerRelayerFee = await txService.calculateGasFee(
-              chainId,
-              routerRelayerFeeAsset,
-              relayerFeeAssetDecimal,
-              "removeLiquidity",
-              isRouterContract,
-              chainData,
-              requestContext,
-            );
-            const signature = await signRouterRemoveLiquidityTransactionPayload(
-              amount,
-              assetId,
-              routerRelayerFeeAsset,
-              routerRelayerFee.toString(),
-              chainId,
-              wallet,
-            );
+            // hardcode to only use router signer
             result = await contractWriter.removeLiquidityRouterContract(
               chainId,
               amount,
               assetId,
               routerAddress,
-              signature,
-              routerRelayerFeeAsset,
-              routerRelayerFee.toString(),
-              true,
+              "0x",
+              constants.AddressZero,
+              "0",
+              false,
               requestContext,
             );
           } else {