+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Next Release
 
 - [utils] Implement health endpoint call in FallbackSubgraph
+- [router] Improve logging, add waiting for safe confirmations when handling receiver cancelled
+
+## 0.1.17
+
+- [router] Disable removeLiquidity through relayers
 
 ## 0.1.16